Transition to proper unit tests for vulnerabilities class
